# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Gym Equipment Directory API",
    "description": "ジムと設備の検索API（MVP）。\n- アプリは postgresql+asyncpg（非同期）\n- Alembic は postgresql+psycopg2（同期）\n- `gyms.last_verified_at_cached` はトリガーで自動更新\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/equipments": {
      "get": {
        "tags": [
          "equipments",
          "equipments"
        ],
        "summary": "設備マスタ検索（補完用）",
        "description": "設備マスタを name/slug の部分一致で検索します（最大20件）。",
        "operationId": "list_equipments_equipments_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "部分一致（ILIKE風）",
              "title": "Q"
            },
            "description": "部分一致（ILIKE風）"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EquipmentMasterDTO"
                  },
                  "title": "Response List Equipments Equipments Get"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/meta/prefectures": {
      "get": {
        "tags": [
          "meta",
          "meta"
        ],
        "summary": "都道府県一覧（distinct）",
        "description": "登録されているジムの都道府県スラッグを重複なしで返します。空/NULLは除外。",
        "operationId": "list_prefectures_meta_prefectures_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PrefOption"
                  },
                  "type": "array",
                  "title": "Response List Prefectures Meta Prefectures Get"
                }
              }
            }
          },
          "503": {
            "description": "database unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/meta/prefs": {
      "get": {
        "tags": [
          "meta",
          "meta"
        ],
        "summary": "都道府県候補を取得（件数付き）",
        "description": "登録されているジムの都道府県スラッグと件数を返します。空/NULLは除外。",
        "operationId": "list_prefs_meta_prefs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PrefOption"
                  },
                  "type": "array",
                  "title": "Response List Prefs Meta Prefs Get"
                }
              }
            }
          },
          "503": {
            "description": "database unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/meta/cities": {
      "get": {
        "tags": [
          "meta",
          "meta"
        ],
        "summary": "市区町村候補を取得（件数付き）",
        "description": "指定した都道府県スラッグに属する市区町村スラッグと件数を返します。空/NULLは除外。",
        "operationId": "list_cities_meta_cities_get",
        "parameters": [
          {
            "name": "pref",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "pattern": "^[a-z0-9-]+$",
              "description": "都道府県スラッグ（lower）例: chiba",
              "examples": [
                "chiba"
              ],
              "title": "Pref"
            },
            "description": "都道府県スラッグ（lower）例: chiba"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CityOption"
                  },
                  "title": "Response List Cities Meta Cities Get"
                }
              }
            }
          },
          "404": {
            "description": "pref not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "database unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/meta/equipment-categories": {
      "get": {
        "tags": [
          "meta",
          "meta"
        ],
        "summary": "設備カテゴリ一覧（distinct）",
        "description": "登録されている設備カテゴリ名を重複なしで返します。空/NULLは除外。",
        "operationId": "list_equipment_categories_meta_equipment_categories_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CategoryOption"
                  },
                  "type": "array",
                  "title": "Response List Equipment Categories Meta Equipment Categories Get"
                }
              }
            }
          },
          "503": {
            "description": "database unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/meta/equipments": {
      "get": {
        "tags": [
          "meta",
          "meta"
        ],
        "summary": "設備スラッグ一覧（distinct）",
        "description": "登録されている設備スラッグと名称・カテゴリを重複なしで返します。空/NULLは除外。",
        "operationId": "list_equipments_meta_meta_equipments_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EquipmentOption"
                  },
                  "type": "array",
                  "title": "Response List Equipments Meta Meta Equipments Get"
                }
              }
            }
          },
          "503": {
            "description": "database unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CategoryOption": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key",
            "description": "安定キー",
            "example": "chiba"
          },
          "label": {
            "type": "string",
            "title": "Label",
            "description": "表示ラベル",
            "example": "千葉"
          },
          "count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Count",
            "description": "関連件数（pref/city のみ）",
            "example": 128
          },
          "category": {
            "type": "string",
            "title": "Category",
            "description": "旧互換の category キーを残す。",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "key",
          "label",
          "category"
        ],
        "title": "CategoryOption"
      },
      "CityOption": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key",
            "description": "安定キー",
            "example": "chiba"
          },
          "label": {
            "type": "string",
            "title": "Label",
            "description": "表示ラベル",
            "example": "千葉"
          },
          "count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Count",
            "description": "関連件数（pref/city のみ）",
            "example": 128
          },
          "city": {
            "type": "string",
            "title": "City",
            "description": "旧互換の city キーを残す。",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "key",
          "label",
          "city"
        ],
        "title": "CityOption"
      },
...
      "EquipmentMasterDTO": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "設備ID"
          },
          "slug": {
            "type": "string",
            "title": "Slug",
            "description": "スラッグ（lower-case, kebab）"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "表示名"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category",
            "description": "カテゴリ（任意）"
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "name"
        ],
        "title": "EquipmentMasterDTO",
        "example": {
          "category": "free-weights",
          "id": 12,
          "name": "スクワットラック",
          "slug": "squat-rack"
        }
      },
      "EquipmentOption": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key",
            "description": "設備スラッグ",
            "example": "smith-machine"
          },
          "label": {
            "type": "string",
            "title": "Label",
            "description": "表示名",
            "example": "スミスマシン"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category",
            "description": "設備カテゴリ",
            "example": "strength"
          },
          "slug": {
            "type": "string",
            "title": "Slug",
            "description": "旧互換 slug を返す。",
            "readOnly": true
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "旧互換 name を返す。",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "key",
          "label",
          "slug",
          "name"
        ],
        "title": "EquipmentOption"
      },
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "エラーメッセージ"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "examples": [
          {
            "detail": "Not Found"
          }
        ]
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PrefOption": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key",
            "description": "安定キー",
            "example": "chiba"
          },
          "label": {
            "type": "string",
            "title": "Label",
            "description": "表示ラベル",
            "example": "千葉"
          },
          "count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Count",
            "description": "関連件数（pref/city のみ）",
            "example": 128
          },
          "pref": {
            "type": "string",
            "title": "Pref",
            "description": "旧互換の pref キーを残す。",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "key",
          "label",
          "pref"
        ],
        "title": "PrefOption"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local"
    }
  ]
}